Update patch set 6
Patch Set 6:

> Patch Set 6: Verified+1
> 
> @Daisuke: I have added comments in the crosbug. Please let me know if I can address something.

If I read this patch literally, the input current is unlimited (stays in PRE CHARGE zone) all the time if the battery is dead because VBAT is always < VSYSREG_SET. Is that right? If so, why do we think it's safe (AC adapter won't be overloaded)?

Patch-set: 6
Reviewer: Gerrit User 1106949 <1106949@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>