Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This only impacts power on resets right?

Well technically it impacts any reset that causes the EC to be held in reset. There's no impact if the EC is not held in reset. Any pin values that do change will be seen immediately by the EC. In that case, the delay is meaningless.

Patch-set: 1
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>