Try soft reset using different data role
Before sending soft reset when there is no battery and vbus is never low,
we do the followings:

1. Restore previous data role from bbram
2. Change the data role to the previous role
3. Save the reverted data role in bbram
4. Issue soft reset
5. If we come back alive, that means we guessed right.
   Save the previous role in bbram.

If we guess it wrong, the device (may) brownout. It requires physical
reset to try again (using the other role). If we guess it right, we
write back the role (after successful soft reset). So, we can keep
using the right role there on after.

As long as the ac adapter is the same, we'll be able to boot without
physical reset.

BUG=b:36502130,b:35648282

Change-Id: If7a7e046cdf7e6195ffcf3f508fc6ec05383f24a