usb_updater: post a reset during upstart on old versions
If Cr50 does not have support for invalidating the RW header and
restoring it, then upstart should still post a reset. This changes
usb_updater to post a restart during upstart if the RW minor version is
less than 19.

usb_updater -f should only call send_done if the transfer is over usb.

Add '-t' option to enable the invalid rw header after the update.
usb_updater will do it automatically on any update not using '-u' if it
detects that the version is higher than 0.0.18. If you build a dev
image from TOT with a version lower than 0.0.19, you will need to use
-t to activate the image.

BUG=none
BRANCH=none
TEST=manual
	Copy new usb_updater onto the DUT. run 'sync'

	Build two cr50 debug images one with 0.0.19 and one with 0.0.20

	verify update process works

	test_that $DUT_IP -b $BOARD firmware_Cr50Update
	--args="old_release_image=$PATH_TO_PROD_13
	release_image=$PATH_TO_PROD_18 dev_image=$DBG_IMAGE_19"

	Build a firmware image with
	https://review.coreboot.org/#/c/18946/3

	test_that $DUT_IP -b $BOARD firmware_Cr50Update
	--args="old_release_image=$PATH_TO_PROD_18
	release_image=$DBG_IMAGE_19 dev_image=$DBG_IMAGE_20"

Change-Id: I811977de26999b1d26bd2d4126b88b1c55a93931
Signed-off-by: Mary Ruthven <mruthven@chromium.org>