Update patch set 7
Patch Set 7:

(1 comment)

Nick/Aseda/Vincent: I'd really like to move forward with this. This code is copied from cr50 (+ my own modifications/fixes), but it'd be good to give this a good second review anyway.

The end goal is to merge cr50 and common code updater (which is why there is a lot of stubbed functions), and the CLs referenced in the commit message attempt to do that. I don't think we want to wait for these CLs to be reviewed (also, I haven't been able to test this on cr50...).

Anyway, this works well enough on hammer, now that the rest of asynchronous signature verification things are in place.

Patch-set: 7
Reviewer: Gerrit User 1129199 <1129199@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>