BD99955: Do not overwrite the charging current in idle mode
copy from CL:453999

If the charging current is less than the BD9995X's minimum charging
current it is overwritten to BD9995X's minimum charging current. However
in idle mode we write the charging current which is known to the charger
during that time, which can be less than the BD9995X's charging current.
Hence, do not overwrite the charging current in idle mode.

BUG=b:35984679,b:37415851
BRANCH=gru
TEST=Manually tested on Bob. In idle mode charge current is 0mA.

Change-Id: I3c7cfabec7909ee195cd91a07db0e123bec62cec
Signed-off-by: james_chao <james_chao@asus.com>