dingdong: fix HPD-to-PD message translation
With a back-to-back HPD IRQ pulse (500us low width / 2ms high width),
dingdong is not able to send out the same number of PD message that
matches the HPD from a MST monitor. This reason PD message is lost is
because of the defer function where if HPD IRQ is triggered too fast,
the defer function is masked.

This fix uses a counter to track the number of HPQ IRQ being received,
then the HPD IRQ defer function will send out that number of HPD event.

BUG=chromium:711334
BRANCH=firmware-glados-7820.B
TEST=Use oscilliscope to monitor HPD signal going into dingdong
     triggered the proper PD message.
     Display via dingdong is functional.

Change-Id: I3a25a9abf3c7e467a1895188ebfbe509027c15c3
Signed-off-by: Kevin K Wong <kevin.k.wong@intel.com>