BD99955: do not set charge current when charging
Rohm said: If charging current of battery command in address 14h set to
0mA during charging, reference of charge current feedback amp.
(VREF_CHG) is set 0V. Then DCDC may stop switching. (Because of the EA
offset)

So check the charge state before set the charge current. If it is
charging, set the charge current to minimal value.

BUG=b:37415851
BRANCH=gru
TEST=test 'ectool chargecontrol normal/idle/discharge' are working

Change-Id: If1b5d42fb7a692db006f287ee27049ea6f82e425
Signed-off-by: james_chao <james_chao@asus.com>