Update patch set 2
Patch Set 2:

It seems that this config option was added because of back boost issues. Why are we removing it?

Using the `charger` command, it seems that Vbatt is not maintained at ~8656mV when AC is not present. This is changing the existing behaviour.

In general, I'm very cautious of making changes with this charger. It's caused issues in the past.

Patch-set: 2
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>