common/usb_update: Support CMD_UNLOCK_RW even with CONFIG_RWSIG is unset
Starting with
be294d4ef "hammer: Only define required options for RO and RW"
we do not define CONFIG_RWSIG for the RW section. However,
we want CMD_UNLOCK_RW to be usable when EC is in RW section, to
unlock itself in prevision for an update on the next reboot.

BRANCH=none
BUG=b:35587171
TEST=Update hammer using usb_updater2

Change-Id: I7ae001363697bc3e17d4a59b21413d35101228cb