poppy: Improve debounce logic and handle side-band wake signal
First, improve debounce logic to only notify AP on base detection
change.

Then, handle side-band signals. For that purpose, base EC pulses
detection pin for 100 us to signal out of band USB wake (that
can be used to wake system from deep S3). We filter these, and
only consider a single ~100us during a debounce interval to be
a valid out-of-band signal.

It appears 180 as base detection ADC threshold is a little low,
let's move it up to 200.

BRANCH=none
BUG=b:35775062
TEST=Flash hammer and poppy, connect them together. Put poppy in
     deep S3 (powerd_dbus_suspend), hammer can wake poppy on
     key press.
TEST=Change hammer's board_usb_wake to send invalid pulses (too long,
     too short, multiple pulses within 5ms), and see that all those
     pulses are ignored.
TEST=Base is detected, despite the fact that ADC shows 187.

Change-Id: I20bf4151134099f52fe8fe0332524aa1ca5f7d63