Update patch set 13
Patch Set 13:

> Patch Set 13:
> 
> Luis/Yunlian: I'm a bit puzzled by this one. Basically, this throws the compiler warning/error below:
> 
> ec-utils-0.0.1-r3790: armv7a-cros-linux-gnueabi-gcc -std=gnu99 -g -Wall -Werror -Wpointer-arith -Wcast-align -Wcast-qual -Wundef -Wsign-compare -Wredundant-decls -Wmissing-declarations -O0 -I/build/gru/usr/include/libusb-1.0  -I../../include -I../../util  -I../../board/cr50 -I ../../chip/g usb_updater.c   -L/build/gru/usr/lib -lusb-1.0  -L/build/gru/usr/lib -lcrypto  -o usb_updater
> ec-utils-0.0.1-r3790: armv7a-cros-linux-gnueabi-gcc -std=gnu99 -g -Wall -Werror -Wpointer-arith -Wcast-align -Wcast-qual -Wundef -Wsign-compare -Wredundant-decls -Wmissing-declarations -O0 -I/build/gru/usr/include/libusb-1.0  -I../../include -I../../util usb_updater2.c   -L/build/gru/usr/lib -lusb-1.0  -lfmap -o usb_updater2
> ec-utils-0.0.1-r3790: usb_updater2.c: In function 'fetch_header_versions':
> ec-utils-0.0.1-r3790: usb_updater2.c:550:10: error: cast increases required alignment of target type [-Werror=cast-align]
> ec-utils-0.0.1-r3790:         *((const int32_t *)(image+fmaparea->offset));
> ec-utils-0.0.1-r3790:           ^
> ec-utils-0.0.1-r3790: cc1: all warnings being treated as errors
> 
> I'm pretty sure image is 4-byte aligned (output of malloc), and fmaparea->offset is 4-byte aligned too. How do I tell the compiler that's the case? Google searches recommend casting to void* first, which is a big hammer... I could also use a memcpy. Thoughts?

unfortunately, the compiler does not know that image is 4 byte aligned (as far as I can tell) since the type is uint8 *. 
You could do the void * casting first, as it is done a few lines below (4 or 5 lines). Or do the memcpy (as done a few lines above). I think the memcpy looks nicer. For small memcpy, the compiler should inline it and generate an instruction that may be a little more expensive if the pointers are not aligned correctly (which is not the case according to what you tell us).

Patch-set: 13
Reviewer: Gerrit User 1001271 <1001271@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>