g: allow to deep-sleep on either USB PHYs
When (USB-)resuming from deep-sleep, ensure that we restore the same PHY
we were using before sleeping. So when using PHY A, this avoids having a
short disconnection at resume by switching to PHY B then PHY A.

Save the PHY selection register into an unused PMU scratch register.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=b:38160821
TEST=manual: build 'proto2' firmware with CONFIG_LOW_POWER_IDLE defined,
with the chip connected to the host on PHY A, make the host issue a USB
Suspend then resume and see no disconnection.

Change-Id: I7abd5e338e5c688c2dd486293f520049cdfd273b