cr50: Add board ID check
WORK IN PROGRESS!  DO NOT SUBMIT!

Add Board ID fields to the signed image header.

Add Board ID fields to INFO1, and a console command to set them.  The
console command will fail if any of:
  - Board ID is already set
  - Board ID would not allow the current image header

Add a check at load-time.  This check is not currently connected to
anything; it should make the load fail.

Sysinfo now prints the Board ID.

BUG=b:35586335
BRANCH=none
TEST=TBD

Change-Id: I4b62ee8746f51aadbbbb3aded2455e3f44822332
Signed-off-by: Randall Spangler <rspangler@chromium.org>