cr50: Add console and TPM vendor commands to get/set board ID
This patch adds vendor and console commands to read and write the
board ID space in the INFO1 block.

Current image's board ID settings are saved in the image header by the
latest codesigner.

Board ID write attempts are rejected if the board ID space is already
initialized, or if the currently running image will not be allowed to
run with the new board ID space settings.

Error codes are returned to the caller as a single byte value.
Successful read command returns 12 bytes of the board ID space
contents.

The console command always allows to read the board ID value, and
allows to write it if the image was built with debug enabled.

BUG=b:35586335
BRANCH=cr50
TEST=as follows:
   - verified that board ID can be read by any image and set by debug
     images.

   - with the upcoming patches verified the ability to set and read
     board ID values using vendor commands.

Change-Id: I1ead7e460fabf3b35dbcfb99147847c8c4402d34
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: ee545922389739b39cc0ac7e0f0d1dd8c2c67607
Original-Change-Id: I35a3e2db92175a29de8011172b80091065b27414
Original-Signed-off-by: Philip Chen <philipchen@google.com>
Original-Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/522234
Original-Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Original-Reviewed-by: Mary Ruthven <mruthven@chromium.org>