Reef: Wake up by lid
With this change, a device should wake up when the lid crosses
the clamshell-tablet boundary. This gives an OS a chance to enable
the touchpad as a wake-up source.

BUG=b:35585566
BRANCH=none
TEST=Verify Electro wakes up when lid goes from front to back
or from back to front. Verify Electro does not wake up by
keyboard or touchpad when lid is in tablet zone. Verify Electro
wakes up when lid is in clamshell zone.

Change-Id: Ic608e84f0e369fc4ce3a7a8bae0a758d9cc944c5
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>