keyboard: Support for inverting different keyboard column
With a revised keyboard matrix the Refresh key is moving to F2,
which means we need to invert column 3 instead of column 2.

Add support for this by making the define include the column to
invert, and modify the existing chips+boards accordingly.

There is no change in functionality.

BUG=b:36735408
BRANCH=none
TEST=make -j buildall

Change-Id: Icdad34555a32977fddd8d1d479f76d9fedb69d98
Signed-off-by: Duncan Laurie <dlaurie@google.com>