npcx: Allow determining inverted keyboard column dynamically
This change will make it possible to support both keyboard matrix
types with one image, as a transition step to the new matrix.

This particular change can be reverted once the new keyboard
matrix is available on all Eve units.

BUG=b:36735408
BRANCH=none
TEST=make -j buildall; manual testing on Eve EVT to ensure that
column 2 is inverted and all keys work as expected.  Unable to
completely test the behavior of column 3 invert yet.

Change-Id: I450b76526fcdc1353256c217b821ad142f2cf248
Signed-off-by: Duncan Laurie <dlaurie@google.com>