hammer: Make keyboard work at Firmware Screen
First, libpayload expects the keyboard interface index to be 0.

Then, hid_iface_request needs to reply to USB_HID_DT_HID request with the
content of struct usb_hid_descriptor. With current code, the variable
name is generated (and therefore hard to guess), so we create a new set
of macros so that we can use a specific variable name.

Finally, add a note regarding USB HID keyboard boot protocol, to make
sure that we do not accidentally change the report format.

BRANCH=none
BUG=b:36538963
TEST=Keyboard works in FW screen, both trackpad and keyboard still work
     when AP has booted.

Change-Id: Ibea4888385909c9ce3b430464e5805c039d4b9ed