Add 'max' value to ec_status enum.
This value is necessary to keep the compiler from making this enum
a one-byte-long value. As it currently stands, variables of this
type can only hold up to uint8_max. These values are placed in the
ec_host_response.result field, which is a uint16_t. This change is
necessary for host commands that wish to return values between
uint8_max and uint16_max.

BRANCH=none
BUG=none
TEST=make buildall -j

Change-Id: Iacdca51dc6c1de677d2fbb59ad6dd2572d21ea7f