Fizz: Set PB to on when system brownout
This patch stores RESET_FLAG_BROWNOUT on S5->S3 transition and clears
it on S3->S5 transition.

Up on boot, EC checks RESET_FLAG_BROWNOUT. If it's set, PB is set to
on so that the system can recover from power loss.

If it's not set, PB is set to idle.

When reboot command is issued, the flag isn't cleared. Therefore, EC
immediately boots AP.

BUG=b:37536389
BRANCH=none
TEST=Unplug AC adapter when system is in G3. Verify system goes back
to G3 when AC adapter is replugged. Do the same for S0.

Change-Id: Iaa6f930585050fdd3511c711b449dff47525066d
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>