Update patch set 1
Patch Set 1:

We were not able to launch a strago-no-vmtest-pre-cq trybot for your change within 90 minutes.

This problem can happen if the trybot waterfall is very busy, or if there is an infrastructure issue. Please notify the sheriff and mark your change as ready again. If this problem occurs multiple times in a row, please file a bug.

Commit queue documentation: http://www.chromium.org/developers/tree-sheriffs/sheriff-details-chromium-os/commit-queue-overview

Patch-set: 1
CC: Gerrit User 1104224 <1104224@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>