Update patch set 3
Patch Set 3:

We're going to need more functionality for "V1" CCD config.  Do you want *all* of that added to the usb updater?  That is, can we turn the usb updater into the Cr50 equivalent of ectool?  How does that interact with plans to make the usb updater common across Cr50 and Poppy base?

Also, some of the commands for "V1" need to behave differently if sent from USB vs. the AP, because USB is a bigger attack surface (malicious USB peripheral vs. someone actually needs to get root on your device).  Are USB commands from the updater *only* available in a dev build?  Or do we need to worry about them being handled by prod firmware?  Or is there a way for us to lock those commands on Cr50 so they're only accepted from the AP and not from USB?

Should we meet briefly tomorrow to discuss?

Patch-set: 3
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>